+++ client/src/apis/axios.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-export default axios.create({
+export const joyrideAxios = axios.create({
   baseURL: 'http://localhost:3000',
   withCredentials: true,
 });
