+++ client/src/components/common/AuthRoute/index.tsx
@@ -1,10 +1,10 @@
 import { useRecoilValue } from 'recoil';
-import { loggedInState } from 'states/atoms';
+import { isLoggedInState } from 'states/atoms';
 import { Navigate, Outlet } from 'react-router-dom';
 
 const AuthRoute = () => {
-  const loggedIn = useRecoilValue(loggedInState);
-  return loggedIn ? <Outlet /> : <Navigate to="/" />;
+  const isLoggedIn = useRecoilValue(isLoggedInState);
+  return isLoggedIn ? <Outlet /> : <Navigate to="/" />;
 };
 
 export default AuthRoute;
