+++ client/src/components/common/PublicRoute/index.tsx
@@ -1,10 +1,10 @@
 import { useRecoilValue } from 'recoil';
-import { loggedInState } from 'states/atoms';
+import { isLoggedInState } from 'states/atoms';
 import { Navigate, Outlet } from 'react-router-dom';
 
 const PublicRoute = () => {
-  const loggedIn = useRecoilValue(loggedInState);
-  return loggedIn ? <Navigate to="/" /> : <Outlet />;
+  const isLoggedIn = useRecoilValue(isLoggedInState);
+  return isLoggedIn ? <Navigate to="/" /> : <Outlet />;
 };
 
 export default PublicRoute;
