+++ client/src/components/signup/FirstSignupForm/FirstSignupForm.module.scss
@@ -18,6 +18,6 @@
   }
 }
 
-.btn {
+.btns {
   margin-top: 5rem;
 }
