+++ client/src/components/signup/SignupForms/index.tsx
@@ -1,3 +1,4 @@
+import SignupTerms from '../SignupTerms';
 import FirstSignupForm from '../FirstSignupForm';
 import SocialLogin from 'components/login/SocialLogin';
 import { Link } from 'react-router-dom';
@@ -7,20 +8,18 @@ import classNames from 'classnames/bind';
 
 const cn = classNames.bind(styles);
 
-interface SignupFormControllerProps {
-  currentStep: number;
-  goNext: () => void;
-  goPrevious: () => void;
+interface SignupFormsProps {
+  step: number;
+  totalSteps: number;
 }
 
-const SignupFormController = ({
-  currentStep,
-  goNext,
-  goPrevious,
-}: SignupFormControllerProps) => (
+const SignupForms = ({ step, totalSteps }: SignupFormsProps) => (
   <>
-    <div className={cn('page', { expand: currentStep === 1 })}>
-      <FirstSignupForm goNext={goNext} />
+    <div className={cn('page', { expand: step === 1 })}>
+      <SignupTerms />
+    </div>
+    <div className={cn('page', { expand: step === 2 })}>
+      <FirstSignupForm />
       <div className={cn('link')}>
         <Link to="/login" className={cn('login')}>
           이미 계정이 있으신가요?
@@ -28,11 +27,10 @@ const SignupFormController = ({
       </div>
       <SocialLogin />
     </div>
-
-    <div className={cn('page', { expand: currentStep === 2 })}>
-      <SecondSignupForm goNext={goNext} goPrevious={goPrevious} />
+    <div className={cn('page', { expand: step === totalSteps })}>
+      <SecondSignupForm />
     </div>
   </>
 );
 
-export default SignupFormController;
+export default SignupForms;
