+++ client/src/states/atoms.ts
@@ -1,15 +1,10 @@
 import { atom } from 'recoil';
-import AuthAPI from 'apis/AuthAPI';
 
-export const loggedInState = atom<boolean>({
-  key: 'loggedIn',
+export const isLoggedInState = atom<boolean>({
+  key: 'isLoggedIn',
   default: false,
 });
-
-const authAPI = new AuthAPI();
-export const authAPIState = atom({ key: 'authAPI', default: authAPI });
-
-export const toastState = atom<string | null>({
-  key: 'toast',
-  default: null,
+export const toastMessageState = atom<string>({
+  key: 'toastMessage',
+  default: '',
 });
