+++ client/src/states/selectors.ts
@@ -1,13 +1,13 @@
 import { selector } from 'recoil';
-import { loggedInState } from './atoms';
+import { isLoggedInState } from './atoms';
 
 export const userDataState = selector({
   key: 'userData',
   get: async ({ get }) => {
     // TODO: 캐싱 확인
-    const loggedIn = get(loggedInState);
-    if (!loggedIn) return null;
+    const isLoggedIn = get(isLoggedInState);
+    if (!isLoggedIn) return null;
 
-    return await { image: 'src' };
+    return await true;
   },
 });
