+++ client/src/typescript/interfaces.ts
@@ -1,0 +1,5 @@
+
+export interface SignupStepControls {
+  decreaseStep: () => void;
+  increaseStep: () => void;
+}
